feat: add campfire tool for repo management
