+++ app/Cargo.toml
@@ -70,9 +70,6 @@ tracing = { workspace = true }
 image_hasher = { workspace = true }
 git-version = { workspace = true }
 
-toml_edit = { workspace = true }
-paste = { workspace = true }
-
 [dev-dependencies]
 glam = { workspace = true }
 rusty-hook = "^0.11.2"
