+++ campfire/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "campfire"
+version = { workspace = true }
+edition = "2021"
+
+[dependencies]
+anyhow = { workspace = true }
+clap = { workspace = true, features = ["derive"] }
+once_cell = { workspace = true }
+serde_json = { workspace = true }
+
+rustdoc-json = "0.8.0"
+rustdoc-types = "0.20.0"
