+++ campfire/README.md
@@ -1,0 +1,7 @@
+# Campfire
+
+Campfire is a tool for working with this repository. It offers several subtools for common developer operations.
+
+It is implemented in Rust to benefit from the Cargo tooling and easy cross-platform compatibility.
+
+Play around with the CLI to see what it can do.
