+++ crates/audio/Cargo.toml
@@ -42,6 +42,6 @@ rustfft = { version = "6.1.0" }
 [dev-dependencies]
 color-eyre = "0.6.2"
 pretty_assertions = "1.3"
-macroquad = { version = "=0.3.24", default_features = false, features = [] }
+macroquad = { version = "=0.3.24", default-features = false, features = [] }
 lyon = "1.0"
 approx = "0.5"
