+++ docs/src/reference/pipeline.d.ts
@@ -1,5 +1,4 @@
-// This file is generated by the `generate-docs` tool. Please do not manually edit it.
-// pipeline.json
+// This file is generated by `campfire`. Please do not manually edit it.
 export type u32 = number;
 export type f32 = number;
 export type Vec2 = [number, number];
