--- tools/generate-docs/.gitignore
@@ -1,1 +1,0 @@
-/target
