--- tools/generate-docs/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "generate-docs"
-version = { workspace = true }
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-rustdoc-json = "0.8.0"
-rustdoc-types = "0.20.0"
-serde_json = { workspace = true }
-anyhow = { workspace = true }
-once_cell = { workspace = true }
-
-# This tool is an example as a hack to make sure this doesn't show up for `cargo install`
